Implement basic feature matching
With a latch to set the training descriptors.

Using ORB descriptors since they seem to be the hot new thing.

The features match reasonably well under varying lighting and scale
conditions. Trying to use it on the globe is a bit misleading since
features always get detected in the background, and the feature matching
isn't trying to detect matches that are in the same pattern, just
matches by how much the descriptors are the same (I'm assuming).

Really slow on my phone (1 FPS), but I'm making the calls in java so
there's probably room for improvement. There definitely needs to be room
if I'm going to do anything fancy for the globe drawing part.