Implement time-varying field
Well, just reading the field anyway.

The log packing thing didn't seem to work, but
the accuracy is fine for our purposes it seems.

What's not fine is that the floating point accuracy for
texture lookups makes the edges of continents change depending
on which slice of the texture we're on.

Thus, I'll likely still keep the PNG packing code, but unpack
it to 8 individual textures.

Also, might try floating point textures for communication between
layers, so the noise advection/LIC can be smoother.